- gave typeinfo manipulations into its own submodule;
