+++ tests/test_generalize.py
@@ -1,87 +1,5 @@
 from righttyper.righttyper_types import TypeInfo, TypeInfoSet
-from righttyper.righttyper_utils import union_typeset_str
-from typing import Sequence, cast
-
-
-def generalize(
-    samples: Sequence[tuple[TypeInfo, ...]],
-    *,
-    typevars: dict[tuple[TypeInfo, ...], str]|None = None
-) -> list[str]|None:
-    """
-    Processes a sequence of samples observed for function parameters and return values, looking
-    for patterns that can be replaced with type variables.  If no pattern is detected, the
-    union of those types (per union_typeset_str) is built.
-
-    samples: a sequence of tuples with type information. Each type in a tuple corresponds to
-        a parameter (or return) type.
-    typevars: a dictionary from type tuples (indicating a type usage pattern) to variable names.
-    returns: a list of parameter (or return) type annotations.
-    """
-
-    # Ensure all samples are consistent (the same number of arguments)
-    if any(len(t) != len(samples[0]) for t in samples[1:]):
-        return None
-
-    # By transposing the per-argument types, we obtain tuples with all the
-    # various types seen for each argument.
-    transposed = list(zip(*samples))
-
-    def is_homogeneous_generic(types: tuple[TypeInfo, ...]) -> bool:
-        """Whether the set only contains instances of a single, consistent generic type
-           whose arguments are also all TypeInfo.
-        """
-        if not types:
-            return False
-
-        first = types[0]
-
-        return all(
-            t.module == first.module and
-            t.name == first.name and
-            len(t.args) == len(first.args) and
-            all(isinstance(a, TypeInfo) for a in t.args)
-            for t in types[1:]
-        )
-
-    from collections import Counter
-    from typing import Iterator
-
-    def expand_generics(types: tuple[TypeInfo, ...]) -> Iterator[tuple[TypeInfo, ...]]:
-        yield types
-
-        if is_homogeneous_generic(types):
-            for i in range(len(types[0].args)):
-                # cast dropping 'str' is checked by is_homogeneous_generic
-                yield from expand_generics(cast(tuple[TypeInfo, ...], tuple(t.args[i] for t in types)))
-
-    # Count the number of times a type usage pattern occurs, as we only want to generalize
-    # if one occurs more than once (in more than one argument).
-    occurrences: Counter[tuple[TypeInfo, ...]] = Counter()
-    for types in transposed:
-        occurrences.update([s for s in expand_generics(types)])
-
-    if typevars is None:
-        typevars = {}
-
-    # Rebuild the argument list, defining and replacing type patterns with a type variable.
-    def rebuild(types: tuple[TypeInfo, ...]) -> TypeInfo|str:
-        if is_homogeneous_generic(types):
-            args = tuple(
-                rebuild(cast(tuple[TypeInfo, ...], tuple(t.args[i] for t in types)))
-                for i in range(len(types[0].args))
-            )
-
-            return types[0].replace(args=args)
-
-        if occurrences[types] > 1:
-            if types not in typevars:
-                typevars[types] = f"T{len(typevars)+1}"
-            return typevars[types]
-
-        return union_typeset_str(TypeInfoSet(types))
-
-    return [str(rebuild(types)) for types in transposed]
+from righttyper.typeinfo import generalize, union_typeset_str
 
 
 def ti(name: str, **kwargs) -> TypeInfo:
